requirements.txt
aiosignal==1.3.1
alembic==1.13.1
appdirs==1.4.4
asttokens==2.4.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.0
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
deprecation==2.1.0
docker-pycreds==0.4.0
executing==2.0.1
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2023.12.0
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
idna==3.6
igraph==0.11.3
intel-openmp==2024.0.2
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.6.0
kiwisolver==1.4.5
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mkl==2024.0.0
mkl-include==2024.0.0
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
optuna==3.5.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-igraph==0.11.3
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
ray==2.9.0
rdkit==2022.9.2
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
stack-data==0.6.3
sympy==1.12
tbb==2021.11.0
tensorboardX==2.6.2.2
texttable==1.7.0
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wandb==0.16.1
wcwidth==0.2.12
